GEODE-1279: Rename Bug42055DUnitTest as BucketRegionSizeWithOverflowRegressionTest
The test was apparently unfinished, so I implemented the rest of it. It
looks like the bug still exists (my assertions fail), so I filed
GEODE-4929 and marked the test with Ignore("GEODE-4929"). We need to do
some more work to determine if the bug is really a bug and then fix
this up.