GEODE-4928 DistributedLockService doesn't work as expected while the dlock grantor is initialized
Moved setting of the in-process state to a point before we launch the
background thread to do the cleanup.  Added in-line cleanup with a warning
message if background execution is rejected and we aren't shutting down.

If we _are_ shutting down the in-process state is left being set to true
so that lockers won't get a spurious conflict exception from the grantor
and will retry their commits with a new grantor when one is created.