GEODE-3820 CI failure: DataSerializerPropogationDUnitTest.testDataSerializersEventIdVerification
This was actually fixed in 5fe3c76d6b1a595c250d924d0115e20f2491cd2a but the
flakytest categorization wasn't removed from the test.  I also added a bounce()
of one of the VMs in a test that would fail if the test is accidentally run
multiple times.