GEODE-5085 authentication failure when auto-reconnecting
Added DistributionConfig security properties to the props used to reconnect
to the distributed system.

Modified the getSecurityProps implementation to stop sharing the internal
state of the configuration object and to convert values into external form.
Security-log-level, for instance, was a number instead of the corresponding
level name.

Added a security manager to all of the ReconnectDUnitTest test cases.