GEODE-5085 authentication failure when auto-reconnecting
Reverting getSecurityProps() to return the converted internal form of
property values.  The security service expects these to be in this form
and not the external form.  My change caused a handful of security tests
to fail.

Added toSecurityProperties() in Config.java and converted
InternalDistributedSystem.reconnect() to use this and toProperties()
to get the external form of properties in order to build the new
DistributionConfig to feed the reconnecting system.