GEODE-5085 authentication failure when auto-reconnecting
Modified auto-reconnect to gather the security properties from the
DistributionConfig and to use the non-parsed values of properties
instead of the parsed versions.

Modified unit tests to gather the security properties and use the
non-parsed values when deciding whether to create a new distributed system.

Added a security manager to all of the ReconnectDUnitTest test cases.

This closes #1807