GEODE-5155 hang recovering transaction state for crashed server
I modified the test to hit the issue ~ 70% of the time.  The fix is to
check to see if the message was removed due to a memberDeparted event.
When that happens a departureNoticed flag was being set in TXCommitMessage
but the wait loop in the transaction tracker wasn't checking this flag.