GEODE-5198 NPE in DataSerializer registration when forming a client/server connection during handshake
If a holder can't be found do not record supported classes for it.
Absense of the holder, which had just been inserted into the
idsToHolders collection, means that another thread has resolved the
holder into an actual DataSerializer class and has removed the holder
and its supported classes.

This closes #1943