GEODE-5269 CommitConflictException after TransactionInDoubtException
Before sending a client a TransactionInDoubtException that is caused by
a server shutting down we now wait a bit for the server to finish shutting
down.  This allows any locks it held to be released and avoids a
CommitConflictException if the client should immediately try another
transaction with the same key(s).