GEODE-5307 Hang with servers all in waitForPrimaryMember
Precheckin testing showed that the original commit on this branch broke
partitioning because it wasn't checking to see if the primaryElector
was the member creating the bucket.  In that case there is no Profile in
the RegionAdvisor, so the check would allow BucketAdvisor.volunteerForPrimary()
to proceed.  The fix is to check to see if that's the case.