GEODE-5327: Update expected number of gateway senders
CreateRegionCommandDUnit test contained a race condition that caused a
test to fail if the GatewaySenders were created faster than the 100ms
poll rate of Awaitility. We actually expect two GatewaySenders, one for
each server in the cluster.

The wait was updated to use an assertion for better error messages when
the test fails.

[#158341308]

Co-Authored-by: Bradford D. Boyle <bboyle@pivotal.io>