GEODE-5330: only create one temp folder (#2068)
One temp folder will be left after tests complete. This was the behavior
that GfshCommandRule had prior to GEODE-5327. It is caused by the
creation of the temp directory in the constructor of GfshCommandRule. If
the creation happened only in before(), only one temp directory would be
created and cleaned up. However, this would cause some existing DUnit
tests to break, as they use the GfshCommandRule in a way that never
calls before().

The creation of the temp folder was added to before() as part of
GEODE-5327 so that IntelliJ could run DUnit tests multiple times.