GEODE-5349 State-flush operation may exit early allowing for cache inconsistency
Removed the ability for this method to exit without the operation count
falling to zero.  Instead it issues a fatal-level log message, which
translates into a severe-level alert for operators.  This can help tech
support know which server a customer should terminate in order to break
a distributed deadlock.

I also added an info-level message that is issued if a warning/fatal message
has been issued noting that the wait has completed.  This parallels what
we do in ReplyProcessor21 if we've issued a warning that a cache-op response
hasn't been received within the ack-wait-threshold period.