GEODE-1655: CI Failure in AutoConnectionSourceDUnitTest.testClientMembershipListener
The waitForCrash(), waitForJoin(), waitForLeave() methods in this test were
modified to use Awaitility.  When that was done they started throwing
exceptions but the old versions of these methods didn't do that.  They allowed
the caller to perform validation checks if there was a timeout.

In the case of testClientMembershipListener this has caused the test to fail
if a server's departure is interpreted as not being a "crash" but a "leave".
The validation code allows for that but the altered waitForCrash() method
does not.

I've modified the tests to ignore Awaitility timeouts in these methods, which
should make the test stop being flaky.