GEODE-5385: hang trying to create a bucket
We now look for a ForceReattemptException when destroying a partitioned
region.  This prevents a region ID skew that can occur if another node
is still initializing its region and is not yet ready to destroy it.

I've reenabled the PRSanityCheckMessage that watches for skews like this
and reports them.  This used to be enabled by default but somehow was
disabled a long time ago.