GEODE-5383: Check that data propagated earlier
Check that the data is propagated through WAN before checking the
status of the event queues. This test was flaky because the event queues
were being checked before the events had been added to the event queue,
so the queue was changing as it was being checked. By checking that the
data has been propaged to the other cluster first, we can know that all
events have been added.

Co-authored-by: Finn Southerland <fsoutherland@pivotal.io>