GEODE-5393: StateFlushOperation hangs waiting for non-existant operation to complete
Refactored OperationMonitor so that we decide once whether to track
threads and, if so, create a subclass of OperationMonitor that always
tracks threads.  Otherwise we create a non-thread-tracking OperationMonitor.

The problem with the previous implementation was that the log level can
be changed.  If someone changes the level from config, or higher, to
fine level the code would start throwing UnsupportedOperationExceptions
when trying to modify Collections.emptyMap