GEODE-567: Wait for async datastore creation to reach appropriate point
This test appears like it was trying to test what happens if a put
happens when data store creation is at a particular point.

Unfortunately, it failed to do so correctly because it just spun off the
data store creation in separate threads and then did the put. The wait
criterion used was also completely wrong, so the test would always wait
for 1 minute.

Finally, there were some race conditions caused by passing the same
array list to multiple async task.

Changed the code to wait until the data stores are at the given point in
region creation and then do the put.