GEODE-5248: Fixes in GatewayReceiverMBeanBridge
The class `GatewayReceiverMBeanBridge` uses the `MBeanStatsMonitor` to
retrieve the amount of senders currently connected. This statistic
tends to be out of date and heavily depends on the `HostStatSampler`
class to run periodically and update the data. Whenever a receiver is
stopped its stats are also closed and, as such, the
`clientConnectionCount` is not updated anymore, resulting in a wrong
gateway-sender connected count reported.

- Fixed some minor warnings.
- Added tests to verify the accuracy of the current gateway-senders
  connected.
- Renamed `WanCommandListDUnitTest` to `ListGatewaysCommandDUnitTest`.
- Modified the `GatewayReceiverMBeanBridge` to retrieve the current
  client connection count from the underlying acceptor instead of
  the `MBeanStatsMonitor` class.