GEODE-5437: change async region gii to be synchronous (#2219)
When a region uses async disk persistence it used to write its data to disk async but some of the gii meta-data synchronously. This could cause data inconsistency during recovery because it looked like the disk had the entire gii, when it may not have been able to write all the data.

To fix this we now write all data to the disk store during gii synchronously for regions that are persistent even if they are configured async.

DUnit tests have been added for both backup/restore and killing during gii.
Unit tests have been added to cover the product changes.

Also added bounceForcibly to the test framework which enables bounce to do a mean vm kill unlike bounce which always does a nice vm kill.