GEODE-5528: client event listener invoked multiple times for the same transactional operation
TXCommitMessage.combine() was checking to see if a RegionCommit was already
in its collection by using Collection.contains() but RegionCommit doesn't
implement equals() so this check would return false even if the collection
had a RegionCommit for the same Region.

The fix is to check for the region's path instead.