GEODE-5470: Deleting flaky DlockAndTxLockRegressionTest
This test was introduced as part of the fix for GEODE-4928. However,
we've determined that GEODE-4928 is not actually fixed, which is why
this test is sporadically failing.

Since GEODE-4928 actually appears to be trying to introduce new behavior
for dlocks that is a rather involved feature to implement. Leaving that
feature for another time and deleting this test.