GEODE-5531: Fixed Performance Tests
The class `GfshJuniTest` resulted in failures, not related to this PR,
when the new `StressTest` pipeline CI was ran.
The problem is basically that some methods check the singleton
`LogManager` class, and the interleaving execution of those methods can
lead to a race condition for which some loggers are not redirected and,
as such, the tests fail.
I have split `GfshJuniTest` in two different suites to avoid the
problem and also have different classes to verify the different use
cases of Gfsh: `GfshConsoleModeUnitTest` and `GfshHeadlessModeUnitTest`.