GEODE-5505 Cache listener not invoked on a retried destroy() operation
Delay region synchronization operations to allow clients to retry their
operations.  This keeps a server from obtaining the change made by a
client in another server that was only partially distributed.  Without
the delay it is possible for the sync to complete before a client retries
its operation, so that the client's replay encounters the state change
it already made in its previous attempt.

With this change a client's replay will either be on a server that didn't
see the previous attempt, or it will be on a server that did see the
previous attempt & has the operation recorded in its EventTracker.  In
the first case the operation will be performed again and transmitted
to other servers.  In the second case the server will not apply the operation
to its cache but will forward it to other servers.

The delay is based on the server's "maximum time between pings interval",
which defaults to 1 minute.

This closes #2340