GEODE-5590: Containerize AcceptanceTest
AcceptanceTests were run in parallel, but not in containers, in the CI
pipeline. This caused intermittent port conflicts for tests using
default ports. However, the connector tests use database instances in
docker containers, so to avoid docker in docker issues, the connector
tests can't be run in containers in the pipeline.

* Create new test category, DockerComposedTest
* docker-in-docker is avoided by creating a new test task,
acceptanceTestWithDockerCompose for DockerCompsoedTests. Test run
serially and non-containerized for this task.
* Add depenedency for acceptancetest on acceptanceTestWithDockerCompose