GEODE-3780 member is not considered suspect after failing final check
Ensure that a member is in the suspects collection after it fails a
final check.  This allows processSuspectMessage to know if it should
perform the duty of a membership-coordinator and initiate final
checks based on Suspect messages.

I've also done a little bit of refactoring in processSuspectMessage and
have removed commented-out code.