GEODE-5596 Client ends up with destroyed entry after invalidate()
Fixing failing tests.  The behavior in clients has changed now - if
there is no entry or a tombstone entry in the client and it does an
invalidate() we now expect it to create an entry if concurrency-checks
are enabled.

I also changed the AttributesFactory javadoc for
concurrency-checks-enabled because it was saying it is turned off by
default but that is not the case.