GEODE-5637: disable SingleHopClientExecutorWithLoggingIntegrationTest
This test is super flaky and seems to consistently fail on Windows. It's
trying to use SystemErrRule to test logging which doesn't really work
very well. I think the test needs to be rewritten with a Log4J2 Appender
instead of trying to use stdout or stderr. I'm going to disable it with
Ignore until I can get this rewritten.

Discussed with and reviewed by Sai.