GEODE-5596 Client ends up with destroyed entry after invalidate()
Modifying client cache behavior to create an Invalid entry when an application
invokes invalid() and the entry is currently in Destroyed state on the client
but not on servers.

I also changed the AttributesFactory javadoc for
concurrencyChecksEnabled to note that the default value is true, not
false, and added a test for this setting.

this closes #2384