GEODE-5574: unit test coverage for RegionMapDestroy (#2431)
GEODE-5574: improve unit test coverage of RegionMapDestroy

This is a major rewrite of RegionMapDestroyTest. The test was changed enough that I doubt the diffs will be helpful. I changed it to use mocks except for the class under test.
Each test method now has the pattern "given, do, verify".

This unit test now covers 90% of the class (used to be 60%) and the uncovered code is debug logging and what I think is dead code.

I made one small change to the product to use "putEntryIfAbsent" instead of "getEntryMap().putIfAbsent".