GEODE-5741: Modified tests to be independant of environment
Tests using relative paths had depended on running as a user other root.
The tests expceted that file or directpory creation would fail because the user
would not have permission to write to '/'.

A further problem was the several tests had hardcoded '/' in pathnames,
which is incompatible with running on Windows.