GEODE-5741: Modified tests to be independent of environment
Tests using relative paths had depended on running as a user other than root.
The tests expceted that file or directpory creation would fail because the user
would not have permission to write to filesystem root.

A further problem was the several tests had hardcoded '/' in pathnames,
which is incompatible with running on Windows.

* make tests use relative paths
* remove system dependent path separators
* move tests that expect directory creation to fail to StartMemberUtilsTest.