GEODE-5765: Clean up soft references faster in dunit tests
We observed that several tests in AlterRuntimeCommandDUnitTest class
were taking minutes when they should be taking a few seconds at most.
After some debugging, we determined that the JVM itself is struggling to
deal with having 100K+ classes loaded. They are loaded because the test
is repeated starting and stopping the http service, which creates a
classloader and causes soft references to all of the loaded classes to
be cached.

By setting this flag to clear the soft references faster, the test times
go back to normal.