GEODE-5726 Wait for dispatcher to pause
testReplicatedSerialAsyncEventQueueWithPersistenceEnabled_Restart()
starts a cache with an AsyncEventListener, pauses the gateway sender,
puts 1000 entries, stops and restarts the cache with a new listener, and
verifies that all 1000 entry events are delivered to the new listener.

When the other tests in this class pause the gateway sender, they wait
for the event dispatcher to pause before proceeding. This flaky test
neglected to wait. As a result, from time to time a number of events
were delivered to the old listener before the dispatcher paused, and so
were not delivered to the new listener after restart.

The test now waits for the dispatcher to pause before putting the 1000
entries, and so all of the events are delivered to the post-restart
listener.