GEODE-5711: Gfsh prompt for JNDI username and password
Added Intercepter to CreateJndiCommand to prompt for user name and
password.

Modified GfshParseResult to allow update of the user input text.
When the command parsing occurs again on the remote side, the
modified user input will be used to determine function parms.

Added tests to cover new interceptor and modified tests to cover
new funcitonality for GfshExecutionStrategy and
CreateJndiBindingCommand.

Co-authored-by: Scott Jewell <sjewell@pivotal.io>
Co-authored-by: Jianxia Chen <jchen@pivotal.io>
Co-authored-by: Ben Ross <bross@pivotal.io>
Co-authored-by: Rob Day-Reynolds <rdayreynolds@pivotal.io>