GEODE-5741: Modified tests to be independent of environment (#2496)
* GEODE-5741: Modified tests to be independent of environment

Tests using relative paths had depended on running as a user other than root.
The tests expceted that file or directpory creation would fail because the user
would not have permission to write to filesystem root.

A further problem was the several tests had hardcoded '/' in pathnames,
which is incompatible with running on Windows.

* make tests use relative paths
* remove system dependent path separators
* move tests that expect directory creation to fail to StartMemberUtilsTest.
* On entry to testGeodeOnClasspathIsFirst there may (depending test platform) 
be a version of geode-core-*.jar on the system classpath, but at an unknown position
in tthe list. Now, a dummy geode-core-*.jar path is inserted in the system
classpath at a known, but not the first, position.