GEODE-5501: Wait for listener to disconnect
In several tests, a listener calls disconnectFromDS() to disconnect its member when a certain event occurs.

The tests do not wait for the listener to finish disconnecting.

When the tests end, the cleanup also calls disconnectFromDS() to disconnect the same member.

Within disconnectFromDS(), InternalDistributedSystem.disconnect() is called twice, once on the system the test framework knows about, and once on whatever system InternalDistributedSystem.getConnectedInstance() returns.

If two calls to disconnect() are in progress at the same time, one blocks until the other finishes.

If the test ends before the listener finishes, the listener appears to wait until a later test reuses the same VM, at which time it disconnects whatever system it finds there. We do not know why it waits.

The later test fails if the listener's second call to disconnect() occurs after the test has started a member in the same VM.

To prevent the problem, we:
* Wait for the listener to finish before ending the test.
* Remove the listener at the end of the test, so that it does not
respond to events in later tests.

Signed-Off-By: Helena Bales <hbales@pivotal.io>