GEODE-4872: Fixing logging tests in PersistentColocatedPartitionedRegionDUnitTest
This test had a number of tests making assertions about log messages.
Unfortunately, these tests had assertions inside of a try/finally block,
and the finally had a return. Therefore the assertion failures were
simply ignored, and the test was just a test that the callables returned
in 60 seconds.

Fixing this test by removing the bad verify calls and returns in finally
blocks.

Cleaning up the way the tests hooked into the logging system.