GEODE-5811: Get gradle to cache tomcat and jetty installs
We were downloading full installs of tomcat and jetty on every test run.
This occasionally resulted in test failures if the download site was
down. Now we get this installs as gradle dependencies, which will be
cached.

Removing the extra tomcat 7 version - it turns out we only need to test
against one version of tomcat 7.