GEODE-5676: Disconnect system before closing SocketCreatorFactory
The MemberStarterRule was closing SocketCreatorFactory before calling
DistributedSystem.disconnect. In the case of
ClusterConfigLocatorRestartDUnitTest there was a reconnect thread
running in the background that ended up throwing a NullPointerException
if the SocketCreatorFactory was closed. This led to an infinite loop in
the reconnect thread.

We should not be messing with the internal state of Geode until we call
disconnect to stop all of Geode's background threads.

Co-Authored-By: Dale Emery <demery@pivotal.io>