GEODE-5910: Disabling acceptance tests in StressNewTest
Not all acceptance tests can be run in parallel in docker containers, so
disabling them in StressNewTest.

Also, enabling docker for acceptance tests if the user has configured
docker. We were getting hard to diagnose failures when running
repeatAcceptanceTest with docker configured, because the build silently
ignored the docker settings. Now the build will not ignore those
settings, so acceptance tests than *can* run in
docker will, and tests that cannot will consistently fail with a docker
related error.