GEODE-5863: Rewrite CQ test that wasn't testing anything
testMultipleExecuteWithInitialResults was not actually testing anything
because it fired off asynchronous tasks, but failed to see if those
tasks threw an exception.

Before this fix, the test always got an exception.  It installed a
test hook to prevent the CQ from starting. That prevented any events
from being queued up. The test then waited for all events to be received
on the client, which timed out.

Reimplemented the test to test the condition described in the comments,
making sure that calling executeWithInitialResults twice does not lose
events that are queued up already.