GEODE-5897: change create jndi-binding to fail on an exception (#2680)
Create jndi-binding will now fail if an exception occurs during the create.
Also the data-source is now validated by asking it for a connection when the jndi-binding is created.
If a connection can not be obtained, the create fails and the data-source is closed.
Previously if an exception occurred when creating a jndi-binding an info message was logged.
Now, as the docs state, a warning is logged.

The problem was caused by an exception being caught, logged, and not rethrown.
Now the gfsh function will get the thrown exception and pass a failure back
to gfsh. For backwards compatibility, the XML parser still logs a warning if
the jndi-binding can not be created, instead of having it cause the cache not
to start.