GEODE-5927: enhance create jndi-binding to configure pool (#2711)
The create jndi-binding --datasource-config-properties will now check each
name for a "pool." prefix. If found then that property, minus the prefix,
will be added to the properties that the PooledDataSourceFactory gets to
configure the pool. The other datasource config properties are still used
to configure the database.
The help strings on create jndi-binding have been updated.

Co-authored-by: @BenjaminPerryRoss