GEODE-5797: throw CacheClosed in ColocationHelper
The ColocationHelper runs in the background and wrote a suspect string
to the logs during this test when the cache was closed at the wrong
time, causing the test to fail.

The ColocationHelper would check a static variable to get colocated
regions, and if the cache was closed at the wrong point in this process,
the static region would have already been reset by closing the cache,
and the ColocationHelper would write a suspect string.

To fix this test, and the behavior of the ColocationHelper, we first
check if the cache is in the process of being cancelled. If it is, a
CacheClosedException is thrown instead of an IllegalStateException.
CacheClosedExceptions are ignored.

* add unit test for changes to ColocationHelper. Tests that a
CacheClosedException is thrown from getColocatedRegion when the cache is
closed.
* remove wait for regions to be ready as it was never the correct fix
for this bug.
* create the senders before the regions, since that is the correct order

Signed-off-by: Dan Smith <dsmith@pivotal.io>