GEODE-5925 Server shutdown delays election of new primary bucket owners
This PR addresses the problem of AcceptorImpl sending out expensive
profile update messages when it's being stopped during Cache.close().
There is no reason to send these messages if the affected regions are
going to be destroyed since that also sends profile updates
to other members of the cluster.

This closes #2717