GEODE-5960: Add test to verify JDBC gfsh commands load properly
Added a new test to the JDBC connector project to verify that the list
of commands loaded via the scanner and the command marker match the list
of commands found via these two sources. This is to avoid
errors/warnings due to commands which appear in Command Marker files but
don't have available source code.

Co-authored-by: Ben Ross <bross@pivotal.io>
Co-authored-by: Darrel Schneider <dschneider@pivotal.io>