GEODE-5686: Remove LocalizedStrings
Fixing problems found by lgtm.com

There were a lot of logging calls that were improperly converted from
(MessageFormatString, Argument, Exception) to a form that treated the
Exception as part of the String.format() call.

There were a number of errors in format strings or number of arguments.
MessageFormat was more forgiving about this than String.format().

There were a few dangling MessageFormat specs that needed to be
converted to String.format() specs.

This closes #2757