GEODE-5072 Java clients should not access or modify internal regions
I relaxed restrictions to allow clients to modify the Dynamic Regions
and Client Health cache regions.  These regions are also modified to
have constraints on the type of values that can be put into them, though
the servers don't check these constraints - only the client cache does
so.

I modified the cache helper and GatewayReceiverCommand to give this
operation full access to the cache.  This is required for PDX type
propagation.

One of the tests I had to deal with was taking over 20 minutes to run so
I trimmed it down.  It now runs in about half that time.

This closes #2835

commit 98927d1f49221b3cd617337299c799de5470e072
Author: Bruce Schuchardt <bschuchardt@pivotal.io>
Date:   Fri Nov 9 16:07:53 2018 -0800

    GEODE-5072 Java clients should not access or modify internal regions

    Modified the kind of cache available to client/server command processing
    in order to restrict access to internal regions.