GEODE-5971: Refactor configuration persistence for all groups
AlterAsyncEventQueueCommand differs from most configuration commands in
that it does not update the configuration for explicit "groups". Instead
it must scan all group configrations to find the AEQ(s) that have a
configuration change.

For this alternate command flow, the interface UpdateAllConfigurationGroups
was added. CommandExecutor now builds the list of groups for configuration
updating to include all groups is the command implements
UpdateAllConfigurationGroups. Otherwise, the list of groups are those
specified with a command --group option, or "cluster" if there is no
--group option.

- removed the method SingleGfshCommand.updateAllConfigs that was
introduced in a previous implementation of this PR

Co-authored-by: Jens Deppe <jdeppe@pivotal.io>
Co-authored-by: Kenneth Howe <khowe@pivotal.io>