GEODE-6077: Fixing race in test hook
IndexOnEntrySetJUnitTest was using wait/notify in an unsafe way - if
the wait happened before the notify, it would never be interrupted.

For some reason the test wanted to do the test hook in a separate
thread, maybe due to thread locals? Still running the hook in a separate
thread, but using CompletableFuture which will do the right thing and
also report exceptions that happened in the hook.