GEODE-6067: add list data-source gfsh command (#2907)
list data-source uses the cluster config to determine what data sources exist.
It will show the following items for each data-source: name, is it pooled, is it in use, and the url.
Internally, data sources are stored as jndi-bindings. If a jndi-binding exists that could not have been created by create data-source, then list data-source will ignore that jndi-binding.